create animations page to link to each animation page
