+++ package.json
@@ -21,6 +21,7 @@
     "next": "13.2.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "sass": "^1.62.0",
     "typescript": "5.0.3"
   }
 }
