+++ src/app/layout.module.css
@@ -14,4 +14,5 @@
 
 .nav a {
     padding-inline: 8px;
+    font-family: var(--font-sans);
 }
